🚀 Release v2.1.1 (#53)
Markdown's behavior pre-v3 was to automatically append mdx_ to strings
that acted as references to third-party extensions. Starting in Markdown
v3, this behavior has been removed. In v2.1.0 of this package, we
therefore lost the ability to refer to subscript (forcing developers
to switch to mdx_subscript) if referring to the extension by string
(but only in Markdown v3).

This release re-enables behavior seen on older versions of Markdown.

* 📚 Add PyPI links to docs, sources, and issues
* ✨ Enable short module name via entrypoint
* 👷‍♀️ Install pytest-travis-fold on Travis
* 📚 Simplify introduction docs
* ❗️ Update tox and sphinx-rtd-theme
    * tox from 3.5.1 to 3.5.2
    * sphinx-rtd-theme from 0.4.1 to 0.4.2
* 📚 Add spacing to History
* 🚀 Improve Makefile clean target
* 📚 Update links to PyPI
* 📚 Fix docstring spacing for consistency
* 📚 Update CLI markdown command in usage
* 📚 Update Sphinx conf
* 🚀 Bump version from 2.1.0 to 2.1.1
* 📚 Update History
* 👷‍♀️📚🔧 Test documentation build in CI